refactor(workout): split log route into EventWorkoutView and StrengthWorkoutView
Extract two rendering paths from the 914-line log.$workoutId.tsx into
dedicated components. Route file slimmed from 914 to 562 lines; all
state, hooks, and handlers remain in the route orchestrator.

New files:
- src/components/workout/event-workout-view.tsx (113 lines)
- src/components/workout/strength-workout-view.tsx (489 lines)
